+++ activemodel/CHANGELOG.md
@@ -1,3 +1,5 @@
-* No changes.
+*   Deprecate `Validator#setup`. This should be done manually now in the validator's constructor.
+
+    *Nick Sutterer*
 
 Please check [4-0-stable](https://github.com/rails/rails/blob/4-0-stable/activemodel/CHANGELOG.md) for previous changes.